Audio/VKAT: Use a circular buffer when recording data; should make it easier to detect and handle start/stop beacons [build fix].
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@103570 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/ValidationKit/utils/audio/vkatCommon.cpp
@@ -1097,7 +1097,7 @@ static int audioTestRecordTone(PAUDIOTESTIOOPTS pIoOpts, PAUDIOTESTENV pTstEnv, 
 
             if (g_uVerbosity >= 2)
                 RTTestPrintf(g_hTest, RTTESTLVL_ALWAYS, "Test #%RU32: Processed %RU64ms (%zu bytes)\n",
-                             idxTest, PDMAudioPropsBytesToMilli(pMix->pProps, cbRead), cbRead);
+                             idxTest, PDMAudioPropsBytesToMilli(pMix->pProps, (uint32_t)cbRead), cbRead);
 
             RTCircBufReleaseReadBlock(pCircBuf, cbRead);
